level_3/requirements.txt
google-genai
websockets
python-dotenv
fastapi
uvicorn
google-adk
a2a-sdk